fix: function return types. Will liekly need to reshuffle what nodes do vs what the type mapper does.

+++ c_binder_mojo/typing.mojo
@@ -56,7 +56,7 @@ alias NUMERIC_TYPES:Dict[String, String] = {
     "short": "Int16",
     "int": "Int32",
     "long": "Int64",
-    "long long": "Int64",
+    "long long": "Int128",
     "long double": "Float64",
     "float": "Float32",
     "double": "Float64",
